+++ modules/@apostrophecms/rich-text-widget/ui/apos/components/AposRichTextWidgetEditor.vue
@@ -8,7 +8,8 @@
         duration: 300,
         zIndex: 2000,
         animation: 'fade',
-        inertia: true
+        inertia: true,
+        placement: 'bottom'
       }"
       :editor="editor"
     >
@@ -209,13 +210,17 @@ export default {
       return this.moduleOptions.defaultOptions;
     },
     editorOptions() {
-      const activeOptions = Object.assign({}, this.options);
+      let activeOptions = Object.assign({}, this.options);
 
-      activeOptions.styles = this.enhanceStyles(
-        activeOptions.styles?.length
-          ? activeOptions.styles
-          : this.defaultOptions.styles
-      );
+      activeOptions = {
+        ...activeOptions,
+        ...this.enhanceStyles(
+          activeOptions.styles?.length
+            ? activeOptions.styles
+            : this.defaultOptions.styles
+        )
+      };
+      delete activeOptions.styles;
 
       // Allow default options to pass through if `false`
       Object.keys(this.defaultOptions).forEach((option) => {
@@ -228,6 +233,10 @@ export default {
       activeOptions.className = (activeOptions.className !== undefined)
         ? activeOptions.className : this.moduleOptions.className;
 
+      if (activeOptions.toolbar.includes('styles')) {
+        activeOptions.toolbar = activeOptions.toolbar.filter(t => t !== 'styles');
+        activeOptions.toolbar = [ 'nodes', 'marks', ...activeOptions.toolbar ];
+      }
       return activeOptions;
     },
     autofocus() {
@@ -243,7 +252,12 @@ export default {
       // If we don't supply a valid instance of the first style, then
       // the text align control will not work until the user manually
       // applies a style or refreshes the page
-      const defaultStyle = this.editorOptions.styles.find(style => style.def);
+      const defaultStyle =
+        this.editorOptions.nodes.length
+          ? this.editorOptions.nodes.find(style => style.def)
+          : this.editorOptions.marks.length
+            ? this.editorOptions.marks.find(style => style.def)
+            : null;
 
       const _class = defaultStyle.class ? ` class="${defaultStyle.class}"` : '';
       return `<${defaultStyle.tag}${_class}></${defaultStyle.tag}>`;
@@ -541,7 +555,13 @@ export default {
           styles[0].def = true;
         }
       }
-      return styles;
+
+      // Split styles into node and mark selects
+      const result = {
+        nodes: styles.filter(style => style.command === 'setNode'),
+        marks: styles.filter(style => style.command !== 'setNode')
+      };
+      return result;
     },
     localizeStyle(style) {
       style.label = this.$t(style.label);
@@ -555,7 +575,8 @@ export default {
       return (apos.tiptapExtensions || [])
         .map(extension => extension({
           ...this.editorOptions,
-          styles: this.editorOptions.styles.map(this.localizeStyle),
+          nodes: this.editorOptions.nodes.map(this.localizeStyle),
+          marks: this.editorOptions.marks.map(this.localizeStyle),
           types: this.tiptapTypes
         }));
     },
@@ -713,13 +734,16 @@ function traverseNextNode(node) {
 
     .apos-button--rich-text {
       position: relative;
-      width: 24px;
       height: 24px;
-      padding: 0;
+      padding: 0 8px;
       border: none;
       border-radius: var(--a-border-radius);
       background-color: transparent;
       color: var(--a-base-1);
+      &.apos-button--icon-only {
+        width: 24px;
+        padding: 0;
+      }
       &:hover {
         background-color: transparent;
       }
@@ -775,7 +799,7 @@ function traverseNextNode(node) {
     align-items: stretch;
     max-width: 100%;
     height: auto;
-    gap: 4px;
+    gap: 6px;
   }
 
   .apos-rich-text-editor__editor :deep(.ProseMirror) {